Convert examples into scripts
